extract TypeHeader
